SmartMatch/TrainingContainer/requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
tqdm
transformers>=4.40
sentencepiece
safetensors
PyYAML>=6.0
accelerate
